fix(solidity): address ethers-v6 parity review findings
- revert strict-flag change to label rename only (compareChains scope
  unchanged); document the intentional export/deploy exclusion
- drop stale executedLocally field from pr1067-parity-policy.json
- remove completed waffle row from hardhat-3-migration.md order-of-work
- disclose that this PR restores dev's currently-red contracts-slither
  CI gate in slither-triage.md
- move compare-pr1067-parity.test.ts to test/scripts/ and gate it on
  PARITY_BASELINE/PARITY_CANDIDATE via describe.skip so bare yarn test
  no longer throws
